+++ cacheman.go
@@ -3,15 +3,16 @@ package main
 import (
 	"cacheman/local"
 	"cacheman/remote"
+	. "cacheman/shared"
 	"fmt"
 	"net/http"
 )
 
-var Config local.Config
+var Cfg Config
 
 func main() {
-	local.GetConfig(&Config)
-	local.GetMirrorList(&Config)
+	local.GetConfig(&Cfg)
+	local.GetMirrorList(&Cfg)
 	//TODO: handle errors
 	http.HandleFunc("/", HandleReq)
 	http.ListenAndServe(":8080", nil)
@@ -20,18 +21,19 @@ func main() {
 func HandleReq(w http.ResponseWriter, r *http.Request) {
 	//TODO: Sanitize request input
 	//TODO: Handle case where 2+ computers are downloading the same file
-	RequestedLocalPath := Config.CacheDir + "/" + r.URL.Path[1:] //add cachedir to path, to not check in /
+	RequestedLocalPath := Cfg.CacheDir + "/" + r.URL.Path[1:] //add cachedir to path, to not check in /
 	RequestedPath := r.URL.Path[1:]
 	fmt.Printf("File requested: %s\n", RequestedPath)
 
 	RemoteRequired := true
 
 	if local.FileExists(RequestedLocalPath) { //is file cached?
-		RemoteRequired = !local.ServeFile(w, RequestedLocalPath, &Config)
+		RemoteRequired = !local.ServeFile(w, RequestedLocalPath, &Cfg)
 	}
 
 	if RemoteRequired {
-		remote.ServeFile(w, RequestedPath, &Config)
+		local.BuildDirTreeForFile(RequestedLocalPath)
+		remote.ServeFile(w, RequestedPath, &Cfg)
 	}
 
 }
